robot reste debout mais s emballe

+++ code/Robot Balance 2023/src/init.h
@@ -29,8 +29,8 @@ const int MXC6655XA_I2C_REGISTER_LENGTH = 0x0F;
 const int MXC6655XA_I2C_BAUD_RATE = 115200;
 
 struct structInfoMouvement {
-  int inclinaison;           // inclinaison en degré (signé)
-  int vitesse;               // vitesse en degré par sec (signé)
+  int16_t inclinaison;           // inclinaison en degré (signé)
+  int16_t vitesse;               // vitesse en degré par sec (signé)
 };
 
 // ************************************* //
